fixed function and add card check function
